refactor(server): share application bootstrap configuration
Extract global API prefix, security middleware, CORS, validation, health response and Swagger setup into a reusable configureApp helper so production startup and e2e tests use the same application configuration.
